+++ package.json
@@ -2,7 +2,7 @@
   "name": "color-builder",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://ymurenko.github.io/color-builder",
+  "homepage": "https://colorbuilder.dev",
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
